fix: Reorganise and eliminate circular dependencies
It's hard to depend on ../utils/common when ../utils/common depends on
you :(
