+++ src/gasPriceOracle/adapters/polygon.ts
@@ -1,8 +1,9 @@
-import { providers } from "ethers";
+import { providers, utils as ethersUtils } from "ethers";
 import { BaseHTTPAdapter, BaseHTTPAdapterArgs } from "../../priceClient/adapters/baseAdapter";
-import { isDefined, toBNWei } from "../../utils";
-import { GasPriceEstimate, gasPriceError } from "../oracle";
-import { eip1559 } from "./eip1559";
+import { isDefined } from "../../utils/TypeGuards";
+import { GasPriceEstimate } from "../types";
+import { gasPriceError } from "../util";
+import { eip1559 } from "./ethereum";
 
 type Polygon1559GasPrice = {
   maxPriorityFee: number | string;
@@ -23,6 +24,8 @@ type GasStationArgs = BaseHTTPAdapterArgs & {
   host?: string;
 };
 
+// @dev toBNWei() is not imported from ../utils because of a circular dependency loop.
+//      The fix is probably to relocate the function estimateTotalGasRequiredByUnsignedTransaction().
 class PolygonGasStation extends BaseHTTPAdapter {
   readonly chainId: number;
 
@@ -39,17 +42,17 @@ class PolygonGasStation extends BaseHTTPAdapter {
     const gasPrice: Polygon1559GasPrice = (gas as GasStationV2Response)?.[strategy];
     if (!this.isPolygon1559GasPrice(gasPrice)) {
       // @todo: generalise gasPriceError() to accept a reason/cause?
-      gasPriceError("getFeeData()", this.chainId, toBNWei(0));
+      gasPriceError("getFeeData()", this.chainId, ethersUtils.parseUnits("0"));
     }
 
     [gasPrice.maxFee, gasPrice.maxPriorityFee].forEach((gasPrice) => {
       if (Number(gasPrice) < 0) {
-        gasPriceError("getFeeData()", this.chainId, toBNWei(gasPrice, 9));
+        gasPriceError("getFeeData()", this.chainId, ethersUtils.parseUnits(gasPrice.toString(), 9));
       }
     });
 
-    const maxPriorityFeePerGas = toBNWei(gasPrice.maxPriorityFee, 9);
-    const maxFeePerGas = toBNWei(gasPrice.maxFee, 9);
+    const maxPriorityFeePerGas = ethersUtils.parseUnits(gasPrice.maxPriorityFee.toString(), 9);
+    const maxFeePerGas = ethersUtils.parseUnits(gasPrice.maxFee.toString(), 9);
 
     return { maxPriorityFeePerGas, maxFeePerGas };
   }
