+++ src/gasPriceOracle/index.ts
@@ -1,1 +1,2 @@
-export { GasPriceEstimate, getGasPriceEstimate } from "./oracle";
+export { getGasPriceEstimate } from "./oracle";
+export { GasPriceEstimate } from "./types";
